requirements.txt
cffi==1.15.1
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pycparser==2.21
pydub==0.25.1
pyparsing==3.0.9
python-dateutil==2.8.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
